feat: airport resolver

+++ libs/fsairlines-gql/src/lib/types/airport-data/airport.type.ts
@@ -1,36 +1,5 @@
 import { Field, Float, Int, ObjectType } from '@nestjs/graphql';
 
-// const data = {
-//   status: 'SUCCESS',
-//   data: [
-//     {
-//       id: 4434,
-//       name: 'Schwechat',
-//       iata: 'VIE',
-//       icao: 'LOWW',
-//       city: 'Vienna',
-//       country: 'Austria',
-//       la_g: '48',
-//       la_p: '6',
-//       la_s: '37',
-//       la_d: 'N',
-//       lat: 48.1103,
-//       lo_g: '16',
-//       lo_p: '34',
-//       lo_s: '10',
-//       lo_d: 'E',
-//       lon: 16.5697,
-//       altitude: '600',
-//       length: '',
-//       fuel: 1.42,
-//       size: 4,
-//       province: 'Lower Austria',
-//       mag_var: 2,
-//       population: 3384845,
-//     },
-//   ],
-// };
-
 @ObjectType()
 export class Airport {
   @Field(() => Int, { description: 'The ID of the airport in FSAirlines.' })
@@ -84,7 +53,7 @@ export class Airport {
   @Field(() => String, { description: 'The province of the airport.' })
   province: string;
 
-  @Field(() => Int, { description: 'The magnetic variation at the airport.' })
+  @Field(() => Float, { description: 'The magnetic variation at the airport.' })
   mag_var: number;
 
   @Field(() => Int, { description: 'The population of the airport.' })
